--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -30,7 +30,7 @@ jobs:
 
       # Step 3: Installs all project dependencies
       - name: Install dependencies
-        run: npm ci
+        run: npm install
 
       # # Step 4: Runs the test suite
       # - name: Run tests