--- a/src/components/QuotaWarning.tsx
+++ b/src/components/QuotaWarning.tsx
@@ -10,7 +10,6 @@ import {
   Link,
   Typography,
   CircularProgress,
-  Countdown
 } from "@mui/material";
 import {
   Warning as WarningIcon,