… .claude/settings.local.json
… README.md
… package.json
… pnpm-lock.yaml
… src/lib/gmail/batch-processor.ts
… src/lib/gmail/batchProcessor/batch-api-processor.ts
… src/lib/gmail/batchProcessor/batch-processor.ts
… src/lib/gmail/batchProcessor/constants.ts
… src/lib/gmail/batchProcessor/data-persister.ts
… src/lib/gmail/batchProcessor/email-processor-service.ts
… src/lib/gmail/batchProcessor/response-parser.ts
… src/lib/gmail/email-queue.ts
… src/lib/gmail/gmail-client.ts
… src/lib/startup.ts