@yipfram 2025-10-01
Yeah well maybe 10k line edit is a lot...